requirements.txt
upstox-python-sdk
numpy
pandas
python-dotenv
requests

